+++ spec/proxy_fetcher/providers/proxypedia_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require "spec_helper"
+
+describe ProxyFetcher::Providers::Proxypedia do
+  before :all do
+    ProxyFetcher.config.provider = :proxypedia
+  end
+
+  it_behaves_like "a manager"
+end